refactor: Restructure Navbar and Sidebar components
Created a new branch 'Navbar-refactor' to reorganize the code within the Navbar and Sidebar components. This refactor includes:

- Moved Navbar related components (DropdownMenu, MobileToggleSidebar, NavbarApps, UserDropdownMenu) into a 'Navbar' folder within the 'components' directory.
- Created a new 'Sidebar' directory to house the Sidebar component.
- Improved code organization and separation of concerns for better maintainability and readability.
